# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ClientService",
    "version": "1.0"
  },
  "paths": {
    "/api/Power/GetFieldForPeriod": {
      "get": {
        "tags": [
          "Power"
        ],
        "parameters": [
          {
            "name": "field",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Power/GetAllFieldsForPeriod": {
      "get": {
        "tags": [
          "Power"
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Power/AddMeasurement": {
      "post": {
        "tags": [
          "Power"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AllFieldsValue"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AllFieldsValue"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AllFieldsValue"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Power/AddBulkMeasurements": {
      "post": {
        "tags": [
          "Power"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Bulk"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Bulk"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Bulk"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Power/AddFieldToOutput": {
      "get": {
        "tags": [
          "Power"
        ],
        "parameters": [
          {
            "name": "timeStamp",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "field",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "value",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Power/UpdateField": {
      "put": {
        "tags": [
          "Power"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFieldObject"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFieldObject"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFieldObject"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Power/DeleteTimestamp": {
      "delete": {
        "tags": [
          "Power"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeStamp"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeStamp"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TimeStamp"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Power/DeleteRange": {
      "delete": {
        "tags": [
          "Power"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Period"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Period"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Period"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Power/MinFieldForPeriod": {
      "get": {
        "tags": [
          "Power"
        ],
        "parameters": [
          {
            "name": "field",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Power/MinAllFieldsForPeriod": {
      "get": {
        "tags": [
          "Power"
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Power/MaxFieldForPeriod": {
      "get": {
        "tags": [
          "Power"
        ],
        "parameters": [
          {
            "name": "field",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Power/MaxAllFieldsForPeriod": {
      "get": {
        "tags": [
          "Power"
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Power/AvgFieldForPeriod": {
      "get": {
        "tags": [
          "Power"
        ],
        "parameters": [
          {
            "name": "field",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Power/AvgAllFieldsForPeriod": {
      "get": {
        "tags": [
          "Power"
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Power/SumFieldForPeriod": {
      "get": {
        "tags": [
          "Power"
        ],
        "parameters": [
          {
            "name": "field",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Power/SumAllFieldsForPeriod": {
      "get": {
        "tags": [
          "Power"
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "operationId": "GetWeatherForecast",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AllFieldsValue": {
        "type": "object",
        "properties": {
          "time": {
            "type": "string",
            "nullable": true
          },
          "measurement": {
            "type": "string",
            "nullable": true
          },
          "globalActivePower": {
            "type": "string",
            "nullable": true
          },
          "globalReactivePower": {
            "type": "string",
            "nullable": true
          },
          "globalIntensity": {
            "type": "string",
            "nullable": true
          },
          "subMetering1": {
            "type": "string",
            "nullable": true
          },
          "subMetering2": {
            "type": "string",
            "nullable": true
          },
          "subMetering3": {
            "type": "string",
            "nullable": true
          },
          "voltage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Bulk": {
        "type": "object",
        "properties": {
          "value": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AllFieldsValue"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "DateOnly": {
        "type": "object",
        "properties": {
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "day": {
            "type": "integer",
            "format": "int32"
          },
          "dayOfWeek": {
            "$ref": "#/components/schemas/DayOfWeek"
          },
          "dayOfYear": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "dayNumber": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "DayOfWeek": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
      "Period": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "nullable": true
          },
          "to": {
            "type": "string",
            "nullable": true
          },
          "hasTo": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "TimeStamp": {
        "type": "object",
        "properties": {
          "timeStamp_": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateFieldObject": {
        "type": "object",
        "properties": {
          "timeStamp": {
            "type": "string",
            "nullable": true
          },
          "field": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WeatherForecast": {
        "type": "object",
        "properties": {
          "date": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "temperatureC": {
            "type": "integer",
            "format": "int32"
          },
          "temperatureF": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "summary": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}